requirements.txt
# Core Dependencies
python-dotenv>=1.0.0

# PDF Processing (only PyMuPDF/fitz is used)
PyMuPDF>=1.24.0

# Vector Store (PostgreSQL + pgvector)
psycopg2-binary>=2.9.9
pgvector>=0.2.4
numpy>=1.26.0

# HTTP / LLM Integration (NVIDIA, Groq, Gemini APIs via requests)
requests>=2.31.0

# API Server
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.0
pydantic>=2.0.0

# Repo Analyzer — Git & real-time
GitPython>=3.1.40
websockets>=12.0

# AST Parsing — Tree-sitter for structural code analysis
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0
tree-sitter-typescript>=0.21.0
tree-sitter-java>=0.21.0
tree-sitter-go>=0.21.0
tree-sitter-rust>=0.21.0
